pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "0xsuid",email = "45615666+0xsuid@users.noreply.github.com"}
]
license = {text = "GPLv3"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "strawberry-graphql[fastapi] (>=0.256.1,<0.257.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^24.12.12"

